requirements.txt
blinker==1.4
click==8.1.3
cloudconvert==2.1.0
gunicorn==21.2.0
hume==0.4.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-dotenv==1.0.0
requests==2.31.0
requests-unixsocket==0.2.0
sortedcontainers==2.4.0
Werkzeug==2.2.3
blinker==1.4
click==8.1.3
Flask==2.2.3
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pillow==10.2.0
python-dotenv==1.0.0
requests==2.31.0
requests-unixsocket==0.2.0
sortedcontainers==2.4.0
Werkzeug==2.2.3